+++ popup/index.js
@@ -1,2 +1,0 @@
-var appTarget = document.getElementById("app");
-new miniflux.Popup({ target: appTarget });