+++ src/background.js
@@ -1,76 +1,9 @@
-import { getBrowser, openOptions } from './browser';
-import { getConfiguration, isConfigurationComplete } from './configuration';
+import { getBrowser, openOptions } from "./browser";
+import { getConfiguration, isConfigurationComplete } from "./configuration";
 
-import { MinifluxApi } from './miniflux';
+import { LinkdingApi } from "./linkding";
 
 const browser = getBrowser();
-let api = null;
-let configuration = null;
-let hasCompleteConfiguration = false;
-
-// Omnibox
-browser.omnibox.onInputStarted.addListener(async () => {
-  configuration = await getConfiguration();
-  hasCompleteConfiguration = isConfigurationComplete(configuration);
-  const description = hasCompleteConfiguration
-    ? 'Search bookmarks in Miniflux'
-    : '⚠️ Please configure the Miniflux-injector extension first';
-
-  browser.omnibox.setDefaultSuggestion({ description });
-
-  if (hasCompleteConfiguration) {
-    api = new MinifluxApi(configuration);
-  } else {
-    api = null;
-  }
-});
-
-browser.omnibox.onInputChanged.addListener(async (text, suggest) => {
-  if (!api) {
-    return;
-  }
-  let config = getConfiguration();
-
-  api
-    .search(text, { limit: config.resultNum })
-    .then(async (entries) => {
-      const entrySuggestions = await Promise.all(
-        entries.map(async (entry) => ({
-          content: config.toMiniflux
-            ? await api.getMinifluxUrl(entry.id)
-            : entry.url,
-          description: entry.title,
-        }))
-      );
-      await suggest(entrySuggestions);
-    })
-    .catch((error) => {
-      console.error(error);
-    });
-});
-
-browser.omnibox.onInputEntered.addListener((content, disposition) => {
-  if (!hasCompleteConfiguration || !content) {
-    return;
-  }
-
-  const isUrl = /^http(s)?:\/\//.test(content);
-  const url = isUrl
-    ? content
-    : `${configuration.baseUrl}/search?q=${encodeURIComponent(content)}`;
-
-  switch (disposition) {
-    case 'currentTab':
-      browser.tabs.update({ url });
-      break;
-    case 'newForegroundTab':
-      browser.tabs.create({ url });
-      break;
-    case 'newBackgroundTab':
-      browser.tabs.create({ url, active: false });
-      break;
-  }
-});
 
 // Connection to search injection content script
 let portFromCS;
@@ -78,38 +11,37 @@ let portFromCS;
 function connected(p) {
   portFromCS = p;
 
-  // When the content script sends the search term, search on miniflux and
+  // When the content script sends the search term, search on linkding and
   // return results
-  portFromCS.onMessage.addListener(function (m) {
-    if (m.action == 'openOptions') {
+  portFromCS.onMessage.addListener(async function (m) {
+    if (m.action == "openOptions") {
       // Open the add on options if the user clicks on the options link in the
       // injected box
       openOptions();
-    } else if (isConfigurationComplete() == false) {
+    } else if ((await isConfigurationComplete()) == false) {
       portFromCS.postMessage({
         message:
-          'Connection to your Miniflux instance is not configured yet! ' +
+          "Connection to your linkding instance is not configured yet! " +
           "Please configure the extension in the <a class='openOptions'>options</a>.",
       });
     } else {
-      let config = getConfiguration();
-      api = new MinifluxApi(config);
-      // Configuration is complete, execute a search on miniflux
+      let config = await getConfiguration();
+
+      const api = new LinkdingApi(config);
+
+      // Configuration is complete, execute a search on linkding
       api
         .search(m.searchTerm, { limit: config.resultNum })
-        .then(async (entries) => {
-          const entrySuggestions = await Promise.all(
-            entries.map(async (entry) => ({
-              url: config.toMiniflux
-                ? await api.getMinifluxUrl(entry.id)
-                : entry.url,
-              title: entry.title || entry.url,
-              author: entry.feed.title,
-              id: entry.id,
-            }))
-          );
-          await portFromCS.postMessage({
-            results: entrySuggestions,
+        .then((results) => {
+          const bookmarkSuggestions = results.map((bookmark) => ({
+            url: bookmark.url,
+            title: bookmark.title || bookmark.website_title || bookmark.url,
+            description: bookmark.description || bookmark.website_description,
+            tags: bookmark.tag_names,
+            date: bookmark.date_modified,
+          }));
+          portFromCS.postMessage({
+            results: bookmarkSuggestions,
             config: config,
           });
         })
@@ -119,4 +51,5 @@ function connected(p) {
     }
   });
 }
+
 browser.runtime.onConnect.addListener(connected);